revise utils
