+++ core/src/util/optional.proto
@@ -1,0 +1,3 @@
+message OptionalSerializerMessage {
+  optional bytes valueIfPresent = 1;
+}
