+++ core/src/util/optional.ts
@@ -1,3 +1,9 @@
+import {
+  IOptionalSerializerMessage,
+  OptionalSerializerMessage,
+} from "../../generated/proto_compiled";
+import { Serializer } from "./serialization";
+
 export class Optional<T> {
   private constructor(
     readonly isPresent: boolean,
@@ -16,17 +22,66 @@ export class Optional<T> {
     else return other;
   }
 
+  /**
+   * Map the value by f if present, else returning an empty Optional.
+   * @param  f               [description]
+   * @return                [description]
+   */
+  map<U>(f: (value: T) => U): Optional<U> {
+    if (this.isPresent) return Optional.of(f(this.valueIfPresent!));
+    else return Optional.empty();
+  }
+
   toString(): string {
     if (this.isPresent) {
       return `Optional.of(${this.get()})`;
     } else return "Optional.empty()";
   }
 
-  private static emptyInstance = new Optional(false, undefined);
+  private static emptyInstance = new Optional<unknown>(false, undefined);
   static empty<T>(): Optional<T> {
-    return Optional.emptyInstance as unknown as Optional<T>;
+    return Optional.emptyInstance as Optional<T>;
   }
   static of<T>(value: T): Optional<T> {
     return new Optional(true, value);
+  }
+}
+
+export class OptionalSerializer<T> implements Serializer<Optional<T>> {
+  private constructor(private readonly valueSerializer: Serializer<T>) {}
+
+  serialize(value: Optional<T>): Uint8Array {
+    const imessage: IOptionalSerializerMessage = {};
+    if (value.isPresent) {
+      imessage.valueIfPresent = this.valueSerializer.serialize(value.get());
+    }
+    const message = OptionalSerializerMessage.create(imessage);
+    return OptionalSerializerMessage.encode(message).finish();
+  }
+
+  deserialize(message: Uint8Array): Optional<T> {
+    const decoded = OptionalSerializerMessage.decode(message);
+    if (Object.hasOwnProperty.call(decoded, "valueIfPresent")) {
+      return Optional.of(
+        this.valueSerializer.deserialize(decoded.valueIfPresent)
+      );
+    } else return Optional.empty();
+  }
+
+  // Weak in both keys and values.
+  private static cache = new WeakMap<
+    Serializer<unknown>,
+    WeakRef<OptionalSerializer<unknown>>
+  >();
+
+  static getInstance<T>(valueSerializer: Serializer<T>): OptionalSerializer<T> {
+    const existingWeak = OptionalSerializer.cache.get(valueSerializer);
+    if (existingWeak !== undefined) {
+      const existing = existingWeak.deref();
+      if (existing !== undefined) return <OptionalSerializer<T>>existing;
+    }
+    const ret = new OptionalSerializer(valueSerializer);
+    OptionalSerializer.cache.set(valueSerializer, new WeakRef(ret));
+    return ret;
   }
 }
