+++ core/src/util/serialization.proto
@@ -1,7 +1,3 @@
-message CollabIDMessage {
-    repeated string pathToBase = 1;
-}
-
 message ArrayMessage {
     repeated bytes elements = 1;
 }
@@ -28,8 +24,4 @@ message DefaultSerializerMessage {
 message PairSerializerMessage {
   required bytes one = 1;
   required bytes two = 2;
-}
-
-message OptionalSerializerMessage {
-  optional bytes valueIfPresent = 1;
 }
