+++ core/src/util/uid.ts
@@ -2,7 +2,8 @@
  * Return a UID representing the pair (replicaID,
  * replicaUniqueNumber), in string form.
  *
- * These UIDs are similar to [Lamport timestamps](https://en.wikipedia.org/wiki/Lamport_timestamp),
+ * These UIDs are sometimes called **causal dots**.
+ * They are similar to [Lamport timestamps](https://en.wikipedia.org/wiki/Lamport_timestamp),
  * except that the number is a per-replica counter instead
  * of a logical clock.
  */
